Animations and delete resources no needed
